+++ CHANGELOG.md
@@ -30,6 +30,8 @@ All notable changes to this project will be documented in this file.
   - Initial implementation
 - Release
   - Publish a Docker image for core components.
+- Telemetry
+  - Refactor flow enricher
 
 ## [v0.8.0](https://github.com/malbeclabs/doublezero/compare/client/v0.7.1...client/v0.8.0) – 2025-12-02